Image_Dataset_generator/requirements.txt
selenium==3.141.0
requests==2.26.0
Pillow==8.4.0